Trait pallet_whitelist::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type RuntimeCall: IsType<<Self as Config>::RuntimeCall> + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin, PostInfo = PostDispatchInfo> + GetDispatchInfo + FullCodec + TypeInfo + From<Call<Self>> + Parameter;
    type WhitelistOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type DispatchWhitelistedOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type Preimages: QueryPreimage<H = Self::Hashing> + StorePreimage;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type RuntimeCall: IsType<<Self as Config>::RuntimeCall> + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin, PostInfo = PostDispatchInfo> + GetDispatchInfo + FullCodec + TypeInfo + From<Call<Self>> + Parameter

The overarching call type.

type WhitelistOrigin: EnsureOrigin<Self::RuntimeOrigin>

Required origin for whitelisting a call.

type DispatchWhitelistedOrigin: EnsureOrigin<Self::RuntimeOrigin>

Required origin for dispatching whitelisted call with root origin.

type Preimages: QueryPreimage<H = Self::Hashing> + StorePreimage

The handler of pre-images.

type WeightInfo: WeightInfo

The weight information for this pallet.

Object Safety

This trait is not object safe.